// frontend/src/app/api/models/mcp-server-status.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { McpToolList } from '../models/mcp-tool-list';
import { PortInfo } from '../models/port-info';
export interface McpServerStatus {
  asOf: string;
  autostart: boolean;
  commandLine: string;
  cwd: string;
  exitCode?: (number | null);
  key: string;
  lastError?: (string | null);
  logSource: string;
  pid?: (number | null);
  port?: (PortInfo | null);
  recentOutput?: Array<string>;
  restartOnCrash: boolean;
  restarts?: number;
  revision: number;
  startedAt?: (string | null);
  state: 'stopped' | 'starting' | 'running' | 'stopping' | 'crashed';
  stoppedAt?: (string | null);
  title: string;
  tools?: (McpToolList | null);
  url: (string | null);
}